feat: Adding MiraiImage and MiraiIcon
