+++ packages/mirai/lib/src/parsers/mirai_switch/widget/mirai_switch_widget.dart
@@ -4,8 +4,6 @@ import 'package:mirai/mirai.dart';
 import 'package:mirai/src/parsers/mirai_material_color/mirai_material_color.dart';
 import 'package:mirai/src/parsers/mirai_switch/mirai_switch.dart';
 
-part 'mirai_switch_widget_state.dart';
-
 class MiraiSwitchWidget extends StatefulWidget {
   const MiraiSwitchWidget({super.key, required this.model});
 
@@ -15,7 +13,28 @@ class MiraiSwitchWidget extends StatefulWidget {
   State<MiraiSwitchWidget> createState() => _MiraiSwitchUiWidget();
 }
 
-class _MiraiSwitchUiWidget extends _MiraiSwitchWidgetState {
+class _MiraiSwitchUiWidget extends State<MiraiSwitchWidget> {
+  //#region State based properties
+  bool isSelected = false;
+  //#endregion
+
+  //#region State based methods
+  /// Change the value of the switch as the user toggles it.
+  void _changeValue(bool value) {
+    isSelected = value;
+    if (widget.model.onChanged != null) {
+      Mirai.onCallFromJson(widget.model.onChanged, context);
+    }
+    setState(() {});
+  }
+  //#endregion
+
+  @override
+  void initState() {
+    super.initState();
+    isSelected = widget.model.initialValue ?? false;
+  }
+
   @override
   Widget build(BuildContext context) {
     final MiraiSwitch model = widget.model;
@@ -29,6 +48,7 @@ class _MiraiSwitchUiWidget extends _MiraiSwitchWidgetState {
         );
       case MiraiSwitchType.adaptive:
         return _buildAdaptiveSwitch(
+          context,
           model,
           isSelected: isSelected,
           onChanged: _changeValue,
@@ -36,6 +56,7 @@ class _MiraiSwitchUiWidget extends _MiraiSwitchWidgetState {
       case MiraiSwitchType.material:
       default:
         return _buildMaterialSwitch(
+          context,
           model,
           isSelected: isSelected,
           onChanged: _changeValue,
@@ -62,6 +83,7 @@ Widget _buildCupertinoSwitch(
 }
 
 Widget _buildAdaptiveSwitch(
+  BuildContext context,
   MiraiSwitch model, {
   required bool isSelected,
   required void Function(bool value) onChanged,
@@ -85,13 +107,14 @@ Widget _buildAdaptiveSwitch(
     trackOutlineColor:
         MaterialStateProperty.all(model.trackOutlineColor?.parse),
     trackOutlineWidth: MaterialStateProperty.all(model.trackOutlineWidth),
-    thumbIcon: MaterialStateProperty.all(model.thumbIconWidget),
-    inactiveThumbImage: model.inactiveThumbImageWidget,
-    activeThumbImage: model.activeThumbImageWidget,
+    thumbIcon: MaterialStateProperty.all(model.thumbIconWidget(context)),
+    inactiveThumbImage: model.inactiveThumbImageWidget(context),
+    activeThumbImage: model.activeThumbImageWidget(context),
   );
 }
 
 Widget _buildMaterialSwitch(
+  BuildContext context,
   MiraiSwitch model, {
   required bool isSelected,
   required void Function(bool value) onChanged,
@@ -115,8 +138,8 @@ Widget _buildMaterialSwitch(
     trackOutlineColor:
         MaterialStateProperty.all(model.trackOutlineColor?.parse),
     trackOutlineWidth: MaterialStateProperty.all(model.trackOutlineWidth),
-    thumbIcon: MaterialStateProperty.all(model.thumbIconWidget),
-    inactiveThumbImage: model.inactiveThumbImageWidget,
-    activeThumbImage: model.activeThumbImageWidget,
+    thumbIcon: MaterialStateProperty.all(model.thumbIconWidget(context)),
+    inactiveThumbImage: model.inactiveThumbImageWidget(context),
+    activeThumbImage: model.activeThumbImageWidget(context),
   );
 }
